ProductCreationBatch/revmeup_api_scrapping/requirements.txt
asgiref==3.2.10
attrs==20.2.0
Automat==20.2.0
cffi==1.14.3
constantly==15.1.0
cryptography==3.1.1
cssselect==1.1.0
Django==3.1.1
djangorestframework==3.11.1
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
itemadapter==0.1.0
itemloaders==1.0.3
jmespath==0.10.0
lxml==4.5.2
parsel==1.6.0
pipe==1.6.0
popen==0.1.20
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pytz==2020.1
queuelib==1.5.0
Scrapy==2.3.0
scrapy-user-agents==0.1.1
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
Twisted==20.3.0
ua-parser==0.10.0
user-agents==2.2.0
w3lib==1.22.0
zope.interface==5.1.0
